ExaBayes 0.9.4 * per-partition branch length output correct * IMPORTANT: changed parameter linking synax. Use + to link (instead of /) and 1-3 to link partitions 1 through 3 * protein GTR revmat can now be employed, if specified in the param-block (use lnl-verification, sometimes bugs occur) * frequency parameter for aa-model params can be specified in the param block * much refactoring of the startup * start-up more efficient now, uses less memory * fixed broken post-processing tools

+++ src/ProtModel.hpp
@@ -52,9 +52,9 @@ namespace std
 
   template<> struct hash<ProtModel>
   {
-    unsigned int operator() (const ProtModel& rhs) const 
+    size_t operator() (const ProtModel& rhs) const 
     {
-      return std::hash<int>()(int(rhs));
+      return std::hash<size_t>()(size_t(rhs));
     }
   }; 
 }
